+++ built/grammar/GrammarKeywordDefinitionNode.js
@@ -34,10 +34,12 @@ class GrammarKeywordDefinitionNode extends AbstractGrammarDefinitionNode_1.defau
             return topHalf;
         const captures = requiredCellTypeNames
             .map((typeName, index) => {
-            const theType = program.getCellType(typeName); // todo: cleanup
-            if (!theType)
+            const cellTypeDefinition = program.getCellTypeDefinition(typeName); // todo: cleanup
+            if (!cellTypeDefinition)
                 throw new Error(`No ${GrammarConstants_1.GrammarConstants.cellType} ${typeName} found`); // todo: standardize error/capture error at grammar time
-            return `        ${index + 1}: ${(theType.getHighlightScope() || defaultHighlightScope) + "." + theType.getId()}`;
+            return `        ${index + 1}: ${(cellTypeDefinition.getHighlightScope() || defaultHighlightScope) +
+                "." +
+                cellTypeDefinition.getId()}`;
         })
             .join("\n");
         const cellTypesToRegex = cellTypeNames => cellTypeNames.map(cellTypeName => `({{${cellTypeName}}})?`).join(" ?");