+++ built/grammar/GrammarProgram.d.ts
@@ -2,7 +2,7 @@ import TreeNode from "../base/TreeNode";
 import AbstractRuntimeProgramConstructorInterface from "./AbstractRuntimeProgramConstructorInterface";
 import AbstractGrammarDefinitionNode from "./AbstractGrammarDefinitionNode";
 import GrammarKeywordDefinitionNode from "./GrammarKeywordDefinitionNode";
-import GrammarCellTypeNode from "./GrammarCellTypeNode";
+import GrammarCellTypeDefinitionNode from "./GrammarCellTypeDefinitionNode";
 import types from "../types";
 declare class GrammarRootNode extends AbstractGrammarDefinitionNode {
     protected _getDefaultNodeConstructor(): any;
@@ -17,11 +17,13 @@ declare class GrammarProgram extends AbstractGrammarDefinitionNode {
     getTargetExtension(): string;
     getKeywordOrder(): any;
     private _cache_cellTypes;
-    getCellTypes(): {
-        [name: string]: GrammarCellTypeNode;
+    getCellTypeDefinitions(): {
+        [name: string]: GrammarCellTypeDefinitionNode;
     };
-    getCellType(word: string): GrammarCellTypeNode;
-    protected _getCellTypes(): {};
+    getCellTypeDefinition(word: string): GrammarCellTypeDefinitionNode;
+    protected _getCellTypeDefinitions(): {
+        [typeName: string]: GrammarCellTypeDefinitionNode;
+    };
     getProgram(): this;
     getKeywordDefinitions(): GrammarKeywordDefinitionNode[];
     getTheGrammarFilePath(): string;